added authorization for CRUD goals

+++ backend/controllers/goalControllers.js
@@ -3,7 +3,7 @@ const asyncHandler = require('express-async-handler');
 const Goal = require('../models/goalModel.js');
 
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find();
+    const goals = await Goal.find({user:req.user.id});
     res.status(200).json(goals);
 });
 
@@ -14,7 +14,10 @@ const setGoal = asyncHandler(async (req, res) => {
     }
     const goal = req.body.text;
     console.log(goal);
-    await Goal.create({text:goal});
+    await Goal.create({
+        text:goal,
+        user:req.user.id
+    });
     res.status(200).json({ message: "set goals"});
 });
 
@@ -25,7 +28,14 @@ const updateGoal = asyncHandler(async (req, res) => {
         res.status(400);
         throw new Error(`Goal ${req.params.id} not found`);
     }
-    const updateGoal = await Goal.findByIdAndUpdate(req.params.id, req.body,{new:true});
+
+    // check the user who want to update the goal is actually goal owner or not
+    if( goal.user.toString() !== req.user.id ){
+        res.status(401);
+        throw new Error(`user is not authorized`);
+    }
+
+    const updateGoal = await Goal.findByIdAndUpdate(req.params.id, req.body);
     res.status(200).json({message: updateGoal});
 });
 
@@ -35,6 +45,13 @@ const deleteGoal = asyncHandler(async (req, res) => {
         res.status(400);
         throw new Error(`Goal ${req.params.id} not found`);
     }
+
+    if( goal.user.toString() !== req.user.id ){
+        res.status(401);
+        throw new Error(`user is not authorized`);
+    }
+
+
     const deletedGoal = await Goal.findByIdAndDelete(req.params.id);
     res.json({message: deletedGoal});  
 });
